Align TranscriptLoggerMiddleware implementation with C# version
The TranscriptLoggerMiddleware works on the Emulator, but fails when run
on Azure Infrastructure. This is caused by the invalid modification of
the incoming Activity.

The modifications then create invalid data leading to this JSON fragment
for reply activities:

{
    // ...
    "recipient":{
        "properties":{
            "role":{}
        },
        "id":"8d51e790-8bd3-4450-af3d-1741838d1354",
        "role":{}
    },
    // ...
}

The role properties should be set to a string, not an object. This
results in a
http 400 status from the bot framework:

{
  "error": {
    "code": "BadSyntax",
    "message": "Invalid or missing Activity in request"
  }
}